tree: initial commit
This patch introduce a tree implementation where nodes can be inserted
by a parent indicator using addr. A node stores addr (link-local) and
target (the DIO prefix address). Also it adds an api to get a list of
nodes to reach the root named as path. We need that to hold a topology
interpretation inside the root node and calculate all path to each node.